+++ src/main/java/com/example/springCoinMarket/dto/CashInDto.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 public class CashInDto {
     private Integer id;
-    private Integer coinWalletId;
-    private Integer coinId;
+    private Long coinWalletId;
+    private Long coinId;
     private Integer countCoin;
 }
