+++ src/main/java/com/example/springCoinMarket/dto/UserDto.java
@@ -10,5 +10,5 @@ public class UserDto {
     private String name;
     private String password;
     private String email;
-    private Integer walletId;
+    private Long walletId;
 }
