+++ src/main/java/com/example/springCoinMarket/service/CoinWalletService.java
@@ -5,13 +5,13 @@ import com.example.springCoinMarket.dto.CoinWalletDto;
 import java.util.HashMap;
 
 public interface CoinWalletService {
-    HashMap<Integer, CoinWalletDto> getCoinWallets();
+    HashMap<Long, CoinWalletDto> getCoinWallets();
 
-    CoinWalletDto getCoinWallet(Integer id);
+    CoinWalletDto getCoinWallet(Long id);
 
     void addCoinWallet(CoinWalletDto coinWalletDto);
 
-    void deleteCoinWallet(Integer id);
+    void deleteCoinWallet(Long id);
 
     void updateCoinWallet(CoinWalletDto coinWalletDto);
 }
